+++ packages/client/src/route.tsx
@@ -1,0 +1,118 @@
+import React, { createElement, ComponentType, PureComponent } from 'react';
+import { forEach } from 'lodash';
+import { fromEvent } from 'rxjs';
+
+import { classNames } from '@roxie/components';
+
+type Router = {
+  path: string[];
+  component: ComponentType<any>;
+}
+
+const ROUTES: { [key:string]:Router } = {},
+      getRoute = () => window.location.hash.substring(1) || '/',
+      getRoutePath = (route: string) => route.split('/');
+
+function matchPath() {
+  const currentRoute = getRoute();
+
+  // basic route (without params)
+  if (ROUTES[currentRoute]) {
+    return { route: { path: currentRoute }, component: ROUTES[currentRoute].component };
+  }
+
+  // advanced route (with params)
+  const currentPath = getRoutePath(currentRoute),
+        params: any = {};
+  for(let route in ROUTES) {
+    const mapRoute = ROUTES[route];
+
+    let matched = true;
+    forEach(mapRoute.path, (routeName: string, index: number) => {
+      if (routeName !== currentPath[index]) {
+        if (routeName.startsWith(':')) { // detect param value
+          params[routeName.substring(1)] = currentPath[index];
+        } else {
+          matched = false;
+          return;
+        }
+      }
+    });
+
+    if (matched) {
+      return { route: { path: currentRoute, params }, component: mapRoute.component };
+    }
+  }
+
+  // with not found screen
+  if (ROUTES['*']) {
+    return { route: { path: currentRoute }, component: ROUTES['*'].component };
+  }
+
+  // without not found screen
+  return { route: { path: currentRoute }, component: null };
+};
+
+export class RouterOutlet extends PureComponent {
+  state = matchPath();
+  hashChange$ = fromEvent(window, 'hashchange').subscribe(() => this.setState(matchPath()));
+
+  componentDidMount() {
+    (!window.location.hash) && (window.location.hash = '/');
+  }
+
+  componentWillUnmount() {
+    this.hashChange$.unsubscribe();
+  }
+
+  render() {
+    const { route, component } = this.state;
+
+    if (!component) {
+      console.error('Component not found!');
+      return null;
+    }
+
+    return createElement(component, { route, key: route.path });
+  }
+}
+
+export const route = (routePath: string = '/') => (component: any) => {
+  ROUTES[routePath] = { component, path: getRoutePath(routePath) };
+  return component;
+}
+
+type ComponentConfig = {
+  view: ComponentType<any>;
+}
+
+export const component = (config: ComponentConfig) => (PageModel: any): any => {
+  const model = new PageModel();
+
+  return class extends PureComponent {
+    async componentDidMount() {
+      Object.assign(model, this.props)
+      if (model.onInit) {
+        await model.onInit();
+        this.forceUpdate();
+      }
+    }
+
+    render() {
+      return createElement(config.view, { ...this.props, ...model });
+    }
+  }
+}
+
+export type Route = {
+  path: string;
+  params: { [key:string]:string }
+}
+
+export function RouteLink(props: any) {
+  const { to = '', className = '', activeClassName = 'active', text = '', children, ...others } = props,
+        currentRoute = window.location.hash.substring(1) || '/',
+        isActive = to === currentRoute,
+        cls = classNames('nav-link', className, { [activeClassName]: isActive });
+  return <a href={to ? `#${to}` : 'javascript:void(0)'} className={cls} {...others}>{text || children}</a>
+}